+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] - DATE
 
 ### Added
+
+- *File* API now provides a *read* method
+  in order to read a file's bytes. Added
+  corresponding tests.
    
 ### Changed
  
 ### Fixed
+
+- Bug in *test_storage.MockSFTPClient.getfo*.
  
 ## [0.1.0] - 2023-03-06